refactor: 包括的型安全性リファクタリング
- StructuredLLMClient: systemPrompt を String? → SystemPrompt? に統一
- ConversationHistory: deinit で stream 終了、turnCount 修正、toolCall/toolResult イベント追加
- ToolResponse: isError Bool → ToolResultContent enum に移行
- ToolChoice.none → .disabled にリネーム
- chat/chatWithDetails の重複排除
- ModelProfile.knowledgeCutoff → YearMonth 型
- JSONSchema.minimum/maximum → Double?
- LLMDynamicStructured 空モジュール削除
- MediaError Sendable 対応、未使用 associatedtype 削除

+++ Sources/LLMClient/Schema/JSONSchema+Factory.swift
@@ -52,10 +52,10 @@ extension JSONSchema {
     /// ```
     public static func integer(
         description: String? = nil,
-        minimum: Int? = nil,
-        maximum: Int? = nil,
-        exclusiveMinimum: Int? = nil,
-        exclusiveMaximum: Int? = nil
+        minimum: Double? = nil,
+        maximum: Double? = nil,
+        exclusiveMinimum: Double? = nil,
+        exclusiveMaximum: Double? = nil
     ) -> JSONSchema {
         JSONSchema(
             type: .integer,
@@ -80,8 +80,8 @@ extension JSONSchema {
     /// ```
     public static func number(
         description: String? = nil,
-        minimum: Int? = nil,
-        maximum: Int? = nil
+        minimum: Double? = nil,
+        maximum: Double? = nil
     ) -> JSONSchema {
         JSONSchema(
             type: .number,
